+++ src/test/kotlin/com/booleworks/kjobs/executor/PriorityTest.kt
@@ -109,7 +109,7 @@ private fun CoroutineScope.setupApi(jobPrioritizer: JobPrioritizer? = null): Job
     val jobPersistence = HashMapJobPersistence()
     val j1Persistence = HashMapDataPersistence<TestInput, TestResult>(jobPersistence)
     val j2Persistence = HashMapDataPersistence<TestInput, TestResult>(jobPersistence)
-    val testingApi = JobFrameworkTestingMode(defaultInstanceName, jobPersistence, this, false) {
+    val testingApi = JobFrameworkTestingMode(defaultInstanceName, jobPersistence, false) {
         jobPrioritizer?.let { executorConfig { this.jobPrioritizer = jobPrioritizer } }
         addJob("J1", j1Persistence, { _, _ -> ComputationResult.Success(TestResult(42)) }) {
             jobConfig {